pyproject.toml
[project]
name = "hf-agent-gaia-30"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "arxiv>=2.2.0",
    "crawl4ai>=0.6.3",
    "duckduckgo-search>=8.0.2",
    "gradio[oauth]>=5.34.0",
    "huggingface-hub>=0.33.0",
    "ipykernel>=6.29.5",
    "jupyter>=1.1.1",
    "langchain>=0.3.25",
    "langchain-community>=0.3.24",
    "langchain-core>=0.3.65",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.2",
    "langchain-huggingface>=0.2.0",
    "langchain-ollama>=0.3.3",
    "langchain-tavily>=0.2.2",
    "langgraph>=0.4.8",
    "langmem>=0.0.27",
    "matplotlib>=3.10.3",
    "networkx>=3.4.2",
    "numpy>=2.2.6",
    "openpyxl>=3.1.5",
    "pgvector>=0.4.1",
    "pymupdf>=1.26.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scikit-learn>=1.7.0",
    "scipy>=1.15.3",
    "tiktoken>=0.9.0",
    "tqdm>=4.67.1",
    "trafilatura>=2.0.0",
    "wikipedia>=1.4.0",
]
